requirements.txt
django>=1.6
djangorestframework
django-summernote
Pillow
django-debug-toolbar
